+++ src/features/contactForm/useContactForm.ts
@@ -1,51 +1,13 @@
 import { useState } from 'react';
-import type { ChangeEventHandler, KeyboardEventHandler, FormEventHandler } from 'react';
+import { useForm } from 'react-hook-form';
+import { zodResolver } from '@hookform/resolvers/zod';
 
-import { validateEmail } from '@utils/validators';
-
-interface ContactFormData {
-  name: string;
-  email: string;
-  subject: string;
-  message: string;
-}
-
-type ContactFormValidation = {
-  [K in keyof ContactFormData]: boolean;
-};
-
-type ContactFormErrors = {
-  [K in keyof ContactFormData]: string;
-};
-
-interface ContactFormNotification {
-  type: string;
-  message: string;
-}
+import { ContactFormSchema } from '@schemas/contact/contact.schema';
+import type { ContactForm } from '@schemas/contact/contact.schema';
+import type { FormNotification } from '@typings/forms/notification';
 
 export default function useContactForm() {
-  const [contactFormData, setContactFormData] = useState<ContactFormData>({
-    name: '',
-    email: '',
-    subject: '',
-    message: '',
-  });
-
-  const contactFormValidation: ContactFormValidation = {
-    name: true,
-    email: true,
-    subject: true,
-    message: true,
-  };
-
-  const [contactFormErrors, setContactFormErrors] = useState<ContactFormErrors>({
-    name: '',
-    email: '',
-    subject: '',
-    message: '',
-  });
-
-  const [contactFormNotification, setContactFormNotification] = useState<ContactFormNotification>({
+  const [contactFormNotification, setContactFormNotification] = useState<FormNotification>({
     type: '',
     message: '',
   });
@@ -57,78 +19,34 @@ export default function useContactForm() {
     });
   };
 
-  const handleFormChange: ChangeEventHandler<HTMLInputElement | HTMLTextAreaElement> = (e) => {
-    const { name, value } = e.target;
-    setContactFormData({
-      ...contactFormData,
-      [name]: value,
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+    setValue,
+    reset,
+  } = useForm({
+    resolver: zodResolver(ContactFormSchema),
+  });
+
+  const onSubmit = async (data: ContactForm) => {
+    console.log(data); // temp
+
+    setContactFormNotification({
+      type: 'success',
+      message: 'Your message has successfully been sent',
     });
+
+    reset();
   };
 
-  const handleFormKeyDown: KeyboardEventHandler<HTMLFormElement> = (e) => {
-    const target = e.target as HTMLElement;
-    const tag = target.tagName.toLowerCase();
-
-    if (e.key === 'Enter' && !(tag === 'textarea' || tag === 'select' || (tag === 'button' && (target as HTMLButtonElement).type === 'submit'))) {
-      e.preventDefault();
-    }
+  return {
+    register,
+    handleSubmit,
+    onSubmit,
+    errors,
+    setValue,
+    contactFormNotification,
+    resetContactFormNotification,
   };
-
-  const validateFormData = () => {
-    let formErrors = { ...contactFormErrors };
-    let formIsValid = true;
-
-    for (const field in contactFormData) {
-      const key = field as keyof ContactFormData;
-
-      const value = contactFormData[key];
-      const required = contactFormValidation[key];
-
-      if (key === 'email') {
-        if ((!value && required) || !validateEmail(contactFormData.email)) {
-          if (!value && required) {
-            formErrors[key] = `Please enter an email address`;
-          } else {
-            formErrors[key] = 'Please enter a valid email address';
-          }
-          formIsValid = false;
-        } else {
-          formErrors[key] = '';
-        }
-      } else {
-        if (!value && required) {
-          formErrors[key] = `Please enter a ${key}`;
-          formIsValid = false;
-        } else {
-          formErrors[key] = '';
-        }
-      }
-    }
-
-    setContactFormErrors(formErrors);
-
-    return formIsValid;
-  };
-
-  const handleFormSubmit: FormEventHandler<HTMLFormElement> = (e) => {
-    e.preventDefault();
-
-    const isFormValid = validateFormData();
-
-    if (isFormValid === true) {
-      setContactFormData({
-        name: '',
-        email: '',
-        subject: '',
-        message: '',
-      });
-
-      setContactFormNotification({
-        type: 'success',
-        message: 'Your message has successfully been sent',
-      });
-    }
-  };
-
-  return { contactFormData, contactFormErrors, contactFormNotification, resetContactFormNotification, handleFormChange, handleFormKeyDown, handleFormSubmit };
 }